update user duration & frequency

+++ CupCarbon-master/src/user/UserList.java
@@ -98,8 +98,8 @@ public class UserList {
 			user.setPreferredLatency(SelectedUser.getDouble("preferredLatency"));
 			user.setPreferredThroughput(SelectedUser.getDouble("preferredThroughput"));
 			user.setPreferredFrequency(SelectedUser.getDouble("preferredFrequency"));
-			user.startTime = SelectedUser.getInteger("startTime");
-			user.endTime = SelectedUser.getInteger("endTime");
+			user.startTime = SelectedUser.getLong("startTime");
+			user.endTime = SelectedUser.getLong("endTime");
 
 
 			users.add(user);
